* Halo1: Holy fuck. Killed ASSERT. All builds now uses YELO_ASSERT flavors. blam::error API is also stubbed out in non-editor builds * Halo1: Yelo::Assert() changed to support warning usages. Won't throw in such cases * Halo1: Moved some widely used game constants to game\game_configuration.hpp * Halo1: Added game_ticks_t and game_time_t to said file * Halo1: Refactored interface structs to their respective interface\.hpp files * Halo1: Some HS system-change-theorizing * Halo1_CheApe: Fixed engine bug where first_person_weapons_initialize_for_new_map doesn't set weapon_index to NONE * Halo1_CheApe: Added tag_block/data allocation wrappers in support of allocation headers * Halo1_CheApe: Changed FIELDSET_INSERT_POSTPROCESSED to require a field name. It is assumed now that _field_pad instances with non-NULL names are postprocessed fields, not useless padding * Halo1_CheApe: Corrected s_tag_field_set_runtime_data's max count constants * Halo1_CheApe: Implemented the non-naive field_comparison_code system * Halo1_CE: Touched up FS code during The Great YELO_ASSERT. priliminary nullptr, auto, range for, unneeded buffer_size paremeters to sprintf_s and such functions which deduce buffer sizes with template magic. Also noted other changes needed in code * Halo1_CE: HTTPClient's g_request_index is now implemented as std::array. We shouldn't need to do/have the index asserts anymore (performed in debug builds of std::array)

+++ OpenSauce/Halo1/Halo1_CheApe/TagGroups/TagGroups.cpp
@@ -41,6 +41,7 @@ namespace Yelo
 			Memory::WriteRelativeJmp(blam__tag_iterator_new, GET_FUNC_VPTR(TAG_ITERATOR_NEW), true);
 			Memory::WriteRelativeJmp(blam::tag_iterator_next, GET_FUNC_VPTR(TAG_ITERATOR_NEXT), true);
 
+//			Memory::WriteRelativeJmp(blam::tag_data_resize, GET_FUNC_VPTR(TAG_DATA_LOAD), true);
 			Memory::WriteRelativeJmp(blam::tag_data_load, GET_FUNC_VPTR(TAG_DATA_LOAD), true);
 #if PLATFORM_ID != PLATFORM_TOOL
 			Memory::WriteRelativeJmp(blam::tag_data_unload, GET_FUNC_VPTR(TAG_DATA_UNLOAD), true);
@@ -232,7 +233,7 @@ namespace Yelo
 
 			__asm	jmp	FUNCTION
 		}
-		API_FUNC_NAKED bool PLATFORM_API tag_data_resize(tag_data* data, size_t new_size)
+		API_FUNC_NAKED bool PLATFORM_API tag_data_resize(tag_data* data, int32 new_size)
 		{
 			static const uintptr_t FUNCTION = GET_FUNC_PTR(TAG_DATA_RESIZE);
